pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "arxiv>=2.4.1",
    "ddgs>=9.11.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
